// frontend/src/shared/lib/api/model/featureResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureResponseStart } from './featureResponseStart';
import type { FeatureResponseEnd } from './featureResponseEnd';
import type { FeatureResponseStartModifier } from './featureResponseStartModifier';
import type { FeatureResponseEndModifier } from './featureResponseEndModifier';
import type { FeatureResponseDescription } from './featureResponseDescription';
import type { FeatureResponseFeatureId } from './featureResponseFeatureId';
import type { FeatureResponseLigand } from './featureResponseLigand';
import type { FeatureResponseAlternativeSequence } from './featureResponseAlternativeSequence';
import type { FeatureResponseEvidence } from './featureResponseEvidence';

export interface FeatureResponse {
  feature_type: string;
  start?: FeatureResponseStart;
  end?: FeatureResponseEnd;
  start_modifier?: FeatureResponseStartModifier;
  end_modifier?: FeatureResponseEndModifier;
  description?: FeatureResponseDescription;
  feature_id?: FeatureResponseFeatureId;
  ligand?: FeatureResponseLigand;
  alternative_sequence?: FeatureResponseAlternativeSequence;
  evidence?: FeatureResponseEvidence;
}